simple_agent.py 일반채팅용으로 변경
