+++ backend/app/chat/prompts/memory_update.py
@@ -1,0 +1,20 @@
+"""
+메모리/프로필 업데이트 프롬프트
+사용자 메시지에서 프로필 업데이트 정보를 추출
+"""
+
+MEMORY_UPDATE_PROMPT = """
+사용자 메시지에서 프로필 업데이트 정보를 추출하세요.
+
+사용자 메시지: "{message}"
+
+JSON 형태로 응답하세요:
+{{
+    "allergies": ["추가할 알레르기"],
+    "dislikes": ["추가할 비선호 음식"],
+    "goals_kcal": 목표칼로리숫자,
+    "goals_carbs_g": 목표탄수화물그램
+}}
+
+업데이트할 항목만 포함하세요.
+"""
