+++ backend/app/chat/prompts/response_generation.py
@@ -1,0 +1,22 @@
+"""
+최종 응답 생성 프롬프트
+검색 결과를 바탕으로 사용자에게 친근한 답변 생성
+"""
+
+RESPONSE_GENERATION_PROMPT = """
+사용자의 키토 식단 관련 질문에 친근하고 도움이 되는 답변을 해주세요.
+
+사용자 질문: "{message}"
+의도: {intent}
+
+{context}
+
+답변 가이드라인:
+1. 한국어로 자연스럽게 답변
+2. 키토 식단의 특성을 고려한 조언 포함
+3. 구체적이고 실용적인 정보 제공
+4. 검색 결과가 있으면 적극 활용
+5. 200-300자 내외로 간결하게
+
+검색 결과가 없는 경우 일반적인 키토 식단 조언을 제공하세요.
+"""
