+++ backend/app/restaurant/prompts/place_search_improvement.py
@@ -1,0 +1,13 @@
+"""
+장소 검색 키워드 개선 프롬프트
+사용자 요청을 카카오 로컬 검색에 적합한 키워드로 변환
+"""
+
+PLACE_SEARCH_IMPROVEMENT_PROMPT = """
+사용자 요청을 카카오 로컬 검색에 적합한 키워드로 변환하세요.
+
+사용자 요청: "{message}"
+
+키토 친화적인 검색 키워드 (1-3개)를 반환하세요:
+예시: "구이", "샤브샤브", "샐러드", "스테이크", "회"
+"""
